Moved teacher to example.
