+++ example/CMakeLists.txt
@@ -1,0 +1,20 @@
+project(Example LANGUAGES CXX)
+
+file(GLOB_RECURSE HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/include/*.hpp")
+file(GLOB_RECURSE SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp")
+
+add_executable(${PROJECT_NAME}
+  ${HEADERS}
+  ${SOURCES}
+  main.cpp
+)
+
+target_link_libraries(${PROJECT_NAME}
+  PRIVATE
+    Network
+)
+
+target_include_directories(${PROJECT_NAME}
+  PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+)
